+++ packages/web-components/src/components/global-header/components/global-header/global-header.ts
@@ -8,11 +8,12 @@
  */
 
 import { customElement } from 'lit/decorators.js';
-import { settings } from '@carbon-labs/utilities/es/settings/index.js';
+import { settings } from '@carbon-labs/utilities';
 import globalHeader from './src/global-header.template.js';
 
 export { CommonHeader } from './src/components/CommonHeader/CommonHeader.js';
 export { HybridIpaasHeader } from './src/components/HybridIpaasHeader/HybridIpaasHeader.js';
+export { HybridIpaasHeaderReactWrapper } from './src/components/HybridIpaasHeaderReactWrapper/HybridIpaasHeaderReactWrapper.js';
 export { LogoutHeader } from './src/components/LogoutHeader/LogoutHeader.js';
 export { LogoutTile } from './src/components/LogoutTile/LogoutTile.js';