+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/HybridIpaasHeader.ts
@@ -10,7 +10,7 @@
 
 import { css, LitElement, html, unsafeCSS } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
-import { settings } from '@carbon-labs/utilities/es/settings/index.js';
+import { settings } from '@carbon-labs/utilities';
 import {
   CUSTOM_EVENT_NAME,
   CUSTOM_EVENT_DETAIL_REFRESH_OPTIONS,
@@ -54,9 +54,15 @@ export class HybridIpaasHeader extends LitElement {
   @property({ type: Function }) aiCallback: (() => void) | undefined;
   @property({ type: String }) aiCallbackEvent = '';
   @property({ type: Function }) logoutCallback: (() => void) | undefined;
+  @property({ type: String }) logoutCallbackEvent = '';
   @property({ type: Function }) notificationOpenCallback:
     | (() => void)
     | undefined;
+  @property({ type: String }) notificationOpenCallbackEvent = '';
+  @property({ type: Function }) searchCallback: (() => void) | undefined;
+  @property({ type: String }) searchCallbackEvent = '';
+  @property({ type: Function }) searchSubmitCallback: (() => void) | undefined;
+  @property({ type: String }) searchSubmitCallbackEvent = '';
   @property({ type: Boolean }) hasNewNotifications = false;
   @property({ type: Object }) searchConfigs: SearchConfigs | null = null;
   @property({ type: Array })
@@ -142,6 +148,14 @@ export class HybridIpaasHeader extends LitElement {
 
     if (this.logoutCallback) {
       footerLink.onClickHandler = this.logoutCallback;
+    } else if (this.logoutCallbackEvent) {
+      footerLink.onClickHandler = () => {
+        const event = new CustomEvent(this.logoutCallbackEvent, {
+          bubbles: true,
+          cancelable: true,
+        });
+        this.dispatchEvent(event);
+      };
     } else {
       footerLink.href = `${this.basePath}/logout`;
     }
@@ -154,7 +168,7 @@ export class HybridIpaasHeader extends LitElement {
       isEnabled: this.solisSidekickEnabled,
       scriptUrl: SOLIS_CDN_HOSTNAMES[env] + '/sidekick/solis-sidekick.es.js',
       insights_enabled: true,
-      reports_enabled: true,
+      overview_enabled: true,
       chat_enabled: false,
       tell_me_more_enabled: false,
     };
@@ -167,6 +181,7 @@ export class HybridIpaasHeader extends LitElement {
         cdn_hostname: SOLIS_CDN_HOSTNAMES[env],
         deployment_environment: solisDeploymentEnvironment[env] || 'local',
         product_id: 'ipaas',
+        backendProxy: '/hybrid-ipaas/v1/proxies/solis/backend',
       };
     } else {
       return {
@@ -176,6 +191,7 @@ export class HybridIpaasHeader extends LitElement {
         cdn_hostname: SOLIS_CDN_HOSTNAMES[env],
         deployment_environment: solisDeploymentEnvironment[env] || 'local',
         product_id: 'ipaas',
+        backendProxy: '/hybrid-ipaas/v1/proxies/solis/backend',
       };
     }
   }
@@ -251,12 +267,47 @@ export class HybridIpaasHeader extends LitElement {
         onClick: this.notificationOpenCallback,
       };
     }
+    if (this.notificationOpenCallbackEvent) {
+      updatedOptions.notificationConfigs = {
+        onClick: () => {
+          const event = new CustomEvent(this.notificationOpenCallbackEvent, {
+            bubbles: true,
+            cancelable: true,
+          });
+          this.dispatchEvent(event);
+        },
+      };
+    }
 
     if (this.searchConfigs) {
       updatedOptions.searchConfigs = {
         placeholder: this.searchConfigs?.placeholder ?? 'Search',
         callback: this.searchConfigs?.callback,
         submitCallback: this.searchConfigs?.submitCallback,
+      };
+    }
+
+    if (this.searchCallbackEvent) {
+      updatedOptions.searchConfigs = {
+        callback: () => {
+          const event = new CustomEvent(this.searchCallbackEvent, {
+            bubbles: true,
+            cancelable: true,
+          });
+          this.dispatchEvent(event);
+        },
+      };
+    }
+
+    if (this.searchSubmitCallbackEvent) {
+      updatedOptions.searchConfigs = {
+        submitCallback: () => {
+          const event = new CustomEvent(this.searchSubmitCallbackEvent, {
+            bubbles: true,
+            cancelable: true,
+          });
+          this.dispatchEvent(event);
+        },
       };
     }