+++ packages/web-components/src/components/global-header/components/global-header/src/components/Search/Search.ts
@@ -10,7 +10,7 @@
 
 import { LitElement, css, html, unsafeCSS } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
-import { settings } from '@carbon-labs/utilities/es/settings/index.js';
+import { settings } from '@carbon-labs/utilities';
 import '@carbon/web-components/es-custom/components/search/index.js';
 
 import styles from './_index.scss?inline' assert { type: 'css' };