+++ packages/web-components/src/components/global-header/components/global-header/src/components/Switcher/__tests__/Switcher.test.ts
@@ -28,7 +28,7 @@ describe('Switcher Component', () => {
         carbonIcon: 'Checkmark',
       },
       {
-        label: 'Manage environments',
+        label: 'Manage custom environments',
         isHeading: false,
         href: '#',
         id: '',