+++ packages/react/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.152.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.151.0...@carbon-labs/react@0.152.0) (2026-01-06)
+
+
+### Bug Fixes
+
+* correct utilities version in two packages ([#963](https://github.com/carbon-design-system/carbon-labs/issues/963)) ([683b13e](https://github.com/carbon-design-system/carbon-labs/commit/683b13e25a7886940b82e2f416b303d168460505))
+
+
+
+
+
+# [0.151.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.150.0...@carbon-labs/react@0.151.0) (2025-12-16)
+
+
+### Bug Fixes
+
+* Complete carbon utilities import update ([#937](https://github.com/carbon-design-system/carbon-labs/issues/937)) ([d853b7a](https://github.com/carbon-design-system/carbon-labs/commit/d853b7a6edeb474e0dab5206ec139a84a68d3e84))
+
+
+
+
+
 # [0.150.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.149.0...@carbon-labs/react@0.150.0) (2025-12-01)