+++ packages/web-components/CHANGELOG.md
@@ -3,6 +3,104 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.111.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.110.0...@carbon-labs/web-components@0.111.0) (2026-01-13)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.110.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.109.0...@carbon-labs/web-components@0.110.0) (2026-01-13)
+
+
+### Features
+
+* adding a backend prop to talk to the solis backend directly ([#971](https://github.com/carbon-design-system/carbon-labs/issues/971)) ([53f6947](https://github.com/carbon-design-system/carbon-labs/commit/53f69476c48cc27e083ce7986f74c7ed20899073))
+
+
+
+
+
+# [0.109.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.108.0...@carbon-labs/web-components@0.109.0) (2026-01-06)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.108.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.107.0...@carbon-labs/web-components@0.108.0) (2026-01-06)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.107.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.106.0...@carbon-labs/web-components@0.107.0) (2026-01-05)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.106.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.105.0...@carbon-labs/web-components@0.106.0) (2026-01-05)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.105.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.104.0...@carbon-labs/web-components@0.105.0) (2025-12-23)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.104.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.102.0...@carbon-labs/web-components@0.104.0) (2025-12-19)
+
+
+### Bug Fixes
+
+* **global-header:** Exclude react and react-dom from vite build ([#950](https://github.com/carbon-design-system/carbon-labs/issues/950)) ([bdd3c45](https://github.com/carbon-design-system/carbon-labs/commit/bdd3c45b3fa74fe6f69d5975431f1208ae126ab5))
+* **global-header:** Export HybridIpaasHeaderReactWrapper correctly ([#949](https://github.com/carbon-design-system/carbon-labs/issues/949)) ([1ebedee](https://github.com/carbon-design-system/carbon-labs/commit/1ebedeef8c47bbd6bf2e5147e714e4ef215598f9))
+
+
+### Features
+
+* **global-header:** Introduce React wrapper for Hybrid iPaaS Header ([#946](https://github.com/carbon-design-system/carbon-labs/issues/946)) ([a3eea61](https://github.com/carbon-design-system/carbon-labs/commit/a3eea6153109c63450ceb8445ffbcc2fbcc507ba))
+
+
+
+
+
+# [0.103.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.102.0...@carbon-labs/web-components@0.103.0) (2025-12-16)
+
+
+### Features
+
+* **global-header:** Introduce React wrapper for Hybrid iPaaS Header ([#946](https://github.com/carbon-design-system/carbon-labs/issues/946)) ([a3eea61](https://github.com/carbon-design-system/carbon-labs/commit/a3eea6153109c63450ceb8445ffbcc2fbcc507ba))
+
+
+
+
+
+# [0.102.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.101.0...@carbon-labs/web-components@0.102.0) (2025-12-16)
+
+
+### Bug Fixes
+
+* Complete carbon utilities import update ([#937](https://github.com/carbon-design-system/carbon-labs/issues/937)) ([d853b7a](https://github.com/carbon-design-system/carbon-labs/commit/d853b7a6edeb474e0dab5206ec139a84a68d3e84))
+
+
+
+
+
 # [0.101.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.100.0...@carbon-labs/web-components@0.101.0) (2025-12-09)
 
 **Note:** Version bump only for package @carbon-labs/web-components