+++ packages/web-components/src/components/global-header/CHANGELOG.md
@@ -3,6 +3,91 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.38.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.37.0...@carbon-labs/wc-global-header@0.38.0) (2026-01-13)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.37.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.36.0...@carbon-labs/wc-global-header@0.37.0) (2026-01-13)
+
+
+### Features
+
+* adding a backend prop to talk to the solis backend directly ([#971](https://github.com/carbon-design-system/carbon-labs/issues/971)) ([53f6947](https://github.com/carbon-design-system/carbon-labs/commit/53f69476c48cc27e083ce7986f74c7ed20899073))
+
+
+
+
+
+# [0.36.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.35.0...@carbon-labs/wc-global-header@0.36.0) (2026-01-06)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.35.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.34.0...@carbon-labs/wc-global-header@0.35.0) (2026-01-05)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.34.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.33.0...@carbon-labs/wc-global-header@0.34.0) (2026-01-05)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.33.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.32.0...@carbon-labs/wc-global-header@0.33.0) (2025-12-23)
+
+**Note:** Version bump only for package @carbon-labs/wc-global-header
+
+
+
+
+
+# [0.32.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.31.0...@carbon-labs/wc-global-header@0.32.0) (2025-12-19)
+
+
+### Bug Fixes
+
+* **global-header:** Exclude react and react-dom from vite build ([#950](https://github.com/carbon-design-system/carbon-labs/issues/950)) ([bdd3c45](https://github.com/carbon-design-system/carbon-labs/commit/bdd3c45b3fa74fe6f69d5975431f1208ae126ab5))
+* **global-header:** Export HybridIpaasHeaderReactWrapper correctly ([#949](https://github.com/carbon-design-system/carbon-labs/issues/949)) ([1ebedee](https://github.com/carbon-design-system/carbon-labs/commit/1ebedeef8c47bbd6bf2e5147e714e4ef215598f9))
+
+
+
+
+
+# [0.31.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.30.0...@carbon-labs/wc-global-header@0.31.0) (2025-12-16)
+
+
+### Features
+
+* **global-header:** Introduce React wrapper for Hybrid iPaaS Header ([#946](https://github.com/carbon-design-system/carbon-labs/issues/946)) ([a3eea61](https://github.com/carbon-design-system/carbon-labs/commit/a3eea6153109c63450ceb8445ffbcc2fbcc507ba))
+
+
+
+
+
+# [0.30.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.29.0...@carbon-labs/wc-global-header@0.30.0) (2025-12-16)
+
+
+### Bug Fixes
+
+* Complete carbon utilities import update ([#937](https://github.com/carbon-design-system/carbon-labs/issues/937)) ([d853b7a](https://github.com/carbon-design-system/carbon-labs/commit/d853b7a6edeb474e0dab5206ec139a84a68d3e84))
+
+
+
+
+
 # [0.29.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-global-header@0.28.0...@carbon-labs/wc-global-header@0.29.0) (2025-12-09)
 
 **Note:** Version bump only for package @carbon-labs/wc-global-header